Use DaemonSets to Install virtctl
Previously, virtctl was installed by an init-container as part of the
KubeVirt deployment. This resulted in it running on only one of the
nodes in a multi-controller configuration and not installing virtctl
on the other controller.

This change ensures that virtctl is installed by moving the
init-container to a DaemonSet that will run on both controllers and
removing the logic from the KubeVirt deployment.

Test Plan:
PASS: Install AIO-SX and verify the creation of /var/opt/kubevirt dir
      and symlink
PASS: Uninstall AIO-SX and verify cleanup of /var/opt/kubevirt
PASS: Install AIO-DX and verify the creation of /var/opt/kubevirt dir,
      virtctl, and symlink
PASS: Uninstall AIO-DX and verify cleanup of /var/opt/kubevirt
PASS: Install AIO-DX and verify the creation of /var/opt/kubevirt dir,
      virtctl, and symlink

Closes-Bug: 2067104
Change-Id: I66b43b1824c061cb3606410c7a20df17f5ae0949
Signed-off-by: Aronsky, Gleb <Gleb.Aronsky@windriver.com>