Fix support for multiple namespaces in Kubevirt
Helm allows charts to span multiple namespaces. However, the
behavior implemented in the Application Framework was using only
namespaces available on kustomization.yaml files to look for enabled
charts. This strategy breaks the multiple namespace support because
kustomize requires each HelmRelease resource to have a different
namespace to support this scenario.

This behavior was resolved in this review:
https://review.opendev.org/c/starlingx/config/+/922932

The kubevirt application needs to adapt as it makes use of namespace
in the root kustomization.yaml which means that a single namespace
will be used across all underlying resources, including Helm
versions.

The solution consists of removing the kube-system namespace from the
root kustomization.yaml and adding the kubevirt namespace to the
kustomization.yaml that is inside the chart.

The "meta.helm.sh/release-namespace" annotation of the
namespaces also needs to be changed to reference the kubevirt
namespace.

Test Plan:
PASS: build-pkgs -c -p python3-k8sapp-kubevirt
PASS: build-pkgs -c -p stx-kubevirt-app-helm
PASS: Upload, apply, remove and delete Kubevirt
PASS: Update the kubevirt app user overrides for the kubevirt
      namespace: "system helm-override-update kubevirt-app
      kubevirt-app kubevirt --set replicas=3". After running the
      command and reapplying the app, the changes were made
      successfully.
PASS: Update the kubevirt app user overrides for the cdi
      namespace: "system helm-override-update kubevirt-app
      kubevirt-app cdi --set replicas=3". After running the command
      and reapplying the app, the changes were made successfully.

Depends-on: https://review.opendev.org/c/starlingx/config/+/922932

Closes-bug:2071380

Change-Id: I7416e80e071b18d31b2f7d0968246365f0e3e9d1
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>