Remove namespace dependency from Helm chart
The kubevirt-app got stuck and was not removed by
itself after the system application-remove command.

After some analysis, it was noted that if a
"Namespace" resource is added as part of the chart
then it will be part of the release. When FluxCD
is deleting the release it waits for all resources
belonging to that release to be deleted, which in
this case will include the namespace itself. In
turn, the namespace cannot be deleted because it
is waiting for the release, therefore causing this
circular dependency issue.

This commit fixes the bug removing the "Namespace"
manifests from the chart and moving them to the
kustomization tree.

Test Plan:
PASS: build-pkgs && build-image.
PASS: AIO-SX fresh install.
PASS: upload/apply/remove/delete kubevirt-app.
PASS: check if actions ran successfully.

Closes-Bug: 2078953

Co-Authored-By: Edson Dias <edson.dias@windriver.com>

Change-Id: I64a61a24522e1d57f59bd01f3b74767afa1d5bdf
Signed-off-by: Igor Soares <Igor.PiresSoares@windriver.com>