Fix for the alarm 500.200
After uploading and applying the kubevirt-app, the 500.200 alarm
is triggered when checking the health status. This change addresses
the issue by increasing the duration and renewBefore value, thereby
preventing the alarm from being raised.

Test Plan:
PASS: Upload and apply the kubevirt-app, then verify that the
500.200 alarm does not appear after running the health status check.
PASS: Perform single and Multi k8s upgrade on AIO-SX.

Closes-Bug: 2084142

Change-Id: I2241eb4e4a6efa00934bceda04c19385abde39c3
Signed-off-by: Boovan Rajendran <boovan.rajendran@windriver.com>